Add files via upload
 This is Alexander's Homework File for Module 1 Challenge.
This Repository contains an Html that displays a webpage called Horiseon. It has three working links at the top of the page that will direct you too the rest of the document. 
Also in this Repository is a CSS file that holds all the Styling that is used for the Html.
Some other things of note that I have added was changes in color and fixing the links which were not working. 
I also made the Document work on Mobile by using the flex.